DataLabeler (1)/requirements.txt
Fabric
Flask-SQLAlchemy
Flask-WTF
flask-login
flask-migrate
WTForms
coverage
dataclasses-json
pandas
flask-paginate
# ecdsa
# paramiko
# pycrypto
